test(core-models): compare tags so the sort stability test can fail

+++ hax-lib/core-models/core-models/src/testing.rs
@@ -153,11 +153,11 @@ impl crate::clone::Clone for CloneWitness {
 #[derive(Clone, Copy, Debug)]
 pub struct Tagged {
     pub key: u8,
-    pub tag: u8,
+    pub tag: usize,
 }
 
 impl Tagged {
-    pub fn new(key: u8, tag: u8) -> Self {
+    pub fn new(key: u8, tag: usize) -> Self {
         Tagged { key, tag }
     }
 }
@@ -213,7 +213,7 @@ pub fn tagged_vec(
     proptest::collection::vec(0u8..3, len).prop_map(|keys| {
         keys.into_iter()
             .enumerate()
-            .map(|(i, key)| Tagged::new(key, i as u8))
+            .map(|(i, key)| Tagged::new(key, i))
             .collect()
     })
 }
